backend/requirements.txt
flask
ultralytics
opencv-python-headless
numpy
python-dotenv
time
pathlib
flask-cors